SNUDY/requirements.txt
asgiref==3.4.1
Django==3.2.6
mysqlclient==2.0.1
pytz==2021.1
sqlparse==0.4.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
djangorestframework-jwt~=1.11.0
django-rest-authtoken
django-filter
drf-yasg # swagger
django-cors-headers
boto3 # access to aws
django-storages
Pillow
factory-boy
requests~=2.26.0